Invocation tests updated
